Python unicode:
Since vim stores all the buffers, variables, ... in one encoding (&enc)
we don't need to use unicode in python scripts, but strings in this
particular encoding. This commit removes decoding to unicode and adds
atplib.buffers module with readlines() and read() methods which returns
the file content. If the file is in buffer list the corresponding buffer
(which is encoded with &enc) is returned. Otherwise we read file from
the disc and assume that it is in &enc.